+++ .tools/test-discovery-migration.php
@@ -56,6 +56,9 @@ class IPSModule
     protected function RegisterAttributeInteger($n, $v) { $this->defs['@' . $n] = $v; }
     public function ReadAttributeInteger($n)  { return (int)($GLOBALS['ATTR'][$this->InstanceID][$n] ?? $this->defs['@' . $n] ?? 0); }
     public function WriteAttributeInteger($n, $v) { $GLOBALS['ATTR'][$this->InstanceID][$n] = $v; }
+    protected function RegisterAttributeBoolean($n, $v) { $this->defs['@' . $n] = $v; }
+    public function ReadAttributeBoolean($n)  { return (bool)($GLOBALS['ATTR'][$this->InstanceID][$n] ?? $this->defs['@' . $n] ?? false); }
+    public function WriteAttributeBoolean($n, $v) { $GLOBALS['ATTR'][$this->InstanceID][$n] = $v; }
     public function UpdateFormField($f, $p, $v) { $GLOBALS['FORMFIELDS'][$f][$p] = $v; }
     protected function ReloadForm() {}
     protected function RegisterVariableBoolean($n, $c, $p, $pos) {}